prov/efa: extract efa_rdm_pke_rtm.c from rxr_pkt_type_req.c
This patch moved all RTM related functions to efa_rdm_pke_rtm.c
