+++ prov/efa/src/rdm/rxr_pkt_type_req.h
@@ -61,239 +61,4 @@ size_t rxr_pkt_req_max_hdr_size(int pkt_type);
 size_t rxr_pkt_max_hdr_size(void);
 
 size_t rxr_pkt_req_data_offset(struct efa_rdm_pke *pkt_entry);
-
-static inline
-struct rxr_rtm_base_hdr *rxr_get_rtm_base_hdr(void *pkt)
-{
-	return (struct rxr_rtm_base_hdr *)pkt;
-}
-
-static inline
-uint32_t rxr_pkt_msg_id(struct efa_rdm_pke *pkt_entry)
-{
-	struct rxr_rtm_base_hdr *rtm_hdr;
-
-	rtm_hdr = rxr_get_rtm_base_hdr(pkt_entry->wiredata);
-	/* only msg and atomic request has msg_id */
-	assert(rtm_hdr->flags & (RXR_REQ_MSG | RXR_REQ_ATOMIC));
-	return rtm_hdr->msg_id;
-}
-
-size_t rxr_pkt_rtm_total_len(struct efa_rdm_pke *pkt_entry);
-
-static inline
-uint64_t rxr_pkt_rtm_tag(struct efa_rdm_pke *pkt_entry)
-{
-	size_t offset;
-	uint64_t *tagptr;
-
-	/*
-	 * In consideration of performance, this function did not cast header
-	 * into different header types to get tag, but assume tag is always
-	 * the last member of header.
-	 */
-	offset = rxr_pkt_req_base_hdr_size(pkt_entry) - sizeof(uint64_t);
-	tagptr = (uint64_t *)(pkt_entry->wiredata + offset);
-	return *tagptr;
-}
-
-static inline
-void rxr_pkt_rtm_settag(struct efa_rdm_pke *pkt_entry, uint64_t tag)
-{
-	size_t offset;
-	uint64_t *tagptr;
-
-	offset = rxr_pkt_req_base_hdr_size(pkt_entry) - sizeof(uint64_t);
-	/* tag is always the last member */
-	tagptr = (uint64_t *)(pkt_entry->wiredata + offset);
-	*tagptr = tag;
-}
-
-static inline
-struct rxr_dc_eager_rtm_base_hdr *rxr_get_dc_eager_rtm_base_hdr(void *pkt)
-{
-	return (struct rxr_dc_eager_rtm_base_hdr *)pkt;
-}
-
-static inline
-struct rxr_dc_eager_msgrtm_hdr *rxr_get_dc_eager_msgrtm_hdr(void *pkt)
-{
-	return (struct rxr_dc_eager_msgrtm_hdr *)pkt;
-}
-
-static inline
-struct rxr_dc_eager_tagrtm_hdr *rxr_get_dc_eager_tagrtm_hdr(void *pkt)
-{
-	return (struct rxr_dc_eager_tagrtm_hdr *)pkt;
-}
-
-static inline
-struct rxr_medium_rtm_base_hdr *rxr_get_medium_rtm_base_hdr(void *pkt)
-{
-	return (struct rxr_medium_rtm_base_hdr *)pkt;
-}
-
-static inline
-struct rxr_dc_medium_rtm_base_hdr *rxr_get_dc_medium_rtm_base_hdr(void *pkt)
-{
-	return (struct rxr_dc_medium_rtm_base_hdr *)pkt;
-}
-
-static inline
-struct rxr_dc_medium_msgrtm_hdr *rxr_get_dc_medium_msgrtm_hdr(void *pkt)
-{
-	return (struct rxr_dc_medium_msgrtm_hdr *)pkt;
-}
-
-static inline
-struct rxr_dc_medium_tagrtm_hdr *rxr_get_dc_medium_tagrtm_hdr(void *pkt)
-{
-	return (struct rxr_dc_medium_tagrtm_hdr *)pkt;
-}
-
-static inline
-struct rxr_longcts_rtm_base_hdr *rxr_get_longcts_rtm_base_hdr(void *pkt)
-{
-	return (struct rxr_longcts_rtm_base_hdr *)pkt;
-}
-
-static inline
-struct rxr_longread_rtm_base_hdr *rxr_get_longread_rtm_base_hdr(void *pkt)
-{
-	return (struct rxr_longread_rtm_base_hdr *)pkt;
-}
-
-static inline
-struct rxr_runtread_rtm_base_hdr *rxr_get_runtread_rtm_base_hdr(void *pkt)
-{
-	return (struct rxr_runtread_rtm_base_hdr *)pkt;
-}
-
-ssize_t rxr_pkt_init_eager_msgrtm(struct efa_rdm_pke *pkt_entry,
-				  struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_dc_eager_msgrtm(struct efa_rdm_pke *pkt_entry,
-				     struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_eager_tagrtm(struct efa_rdm_pke *pkt_entry,
-				  struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_medium_msgrtm(struct efa_rdm_pke *pkt_entry,
-				   struct efa_rdm_ope *txe,
-				   size_t data_offset,
-				   int data_size);
-
-ssize_t rxr_pkt_init_dc_eager_tagrtm(struct efa_rdm_pke *pkt_entry,
-				     struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_dc_medium_msgrtm(struct efa_rdm_pke *pkt_entry,
-				      struct efa_rdm_ope *txe,
-				      size_t data_offset,
-				      int data_size);
-
-ssize_t rxr_pkt_init_medium_tagrtm(struct efa_rdm_pke *pkt_entry,
-				   struct efa_rdm_ope *txe,
-				   size_t data_offset,
-				   int data_size);
-
-ssize_t rxr_pkt_init_dc_medium_tagrtm(struct efa_rdm_pke *pkt_entry,
-				      struct efa_rdm_ope *txe,
-				      size_t data_offset,
-				      int data_size);
-
-ssize_t rxr_pkt_init_longcts_msgrtm(struct efa_rdm_pke *pkt_entry,
-				    struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_dc_longcts_msgrtm(struct efa_rdm_pke *pkt_entry,
-				       struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_longcts_tagrtm(struct efa_rdm_pke *pkt_entry,
-				    struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_dc_longcts_tagrtm(struct efa_rdm_pke *pkt_entry,
-				       struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_longread_msgrtm(struct efa_rdm_pke *pkt_entry,
-				     struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_longread_tagrtm(struct efa_rdm_pke *pkt_entry,
-				     struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_runtread_msgrtm(struct efa_rdm_pke *pkt_entry,
-				     struct efa_rdm_ope *txe,
-				     size_t data_offset,
-				     int data_size);
-
-ssize_t rxr_pkt_init_runtread_tagrtm(struct efa_rdm_pke *pkt_entry,
-				     struct efa_rdm_ope *txe,
-				     size_t data_offset,
-				     int data_size);
-
-static inline
-void rxr_pkt_handle_eager_rtm_sent(struct efa_rdm_ep *ep,
-				   struct efa_rdm_pke *pkt_entry)
-{
-	/* there is nothing to be done for eager RTM */
-	return;
-}
-
-void rxr_pkt_handle_medium_rtm_sent(struct efa_rdm_ep *ep,
-				    struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_longcts_rtm_sent(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_longread_rtm_sent(struct efa_rdm_ep *ep,
-				      struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_runtread_rtm_sent(struct efa_rdm_ep *ep,
-				      struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_eager_rtm_send_completion(struct efa_rdm_ep *ep,
-					      struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_medium_rtm_send_completion(struct efa_rdm_ep *ep,
-					       struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_longcts_rtm_send_completion(struct efa_rdm_ep *ep,
-					     struct efa_rdm_pke *pkt_entry);
-
-static inline
-void rxr_pkt_handle_longread_rtm_send_completion(struct efa_rdm_ep *ep,
-					     struct efa_rdm_pke *pkt_entry)
-{
-}
-
-void rxr_pkt_handle_runtread_rtm_send_completion(struct efa_rdm_ep *ep,
-						 struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_rtm_update_rxe(struct efa_rdm_pke *pkt_entry,
-				 struct efa_rdm_ope *rxe);
-
-/*         This function is called by both
- *            rxr_pkt_handle_rtm_recv() and
- *            efa_rdm_msg_handle_unexp_match()
- */
-ssize_t rxr_pkt_proc_matched_rtm(struct efa_rdm_ep *ep,
-				 struct efa_rdm_ope *rxe,
-				 struct efa_rdm_pke *pkt_entry);
-
-ssize_t rxr_pkt_proc_rtm_rta(struct efa_rdm_ep *ep,
-			     struct efa_rdm_pke *pkt_entry);
-/*
- *         This function handles zero-copy receives that do not require ordering
- */
-void rxr_pkt_handle_zcpy_recv(struct efa_rdm_ep *ep,
-			      struct efa_rdm_pke *pkt_entry);
-/*
- *         This function is shared by all RTM packet types which handle
- *         reordering
- */
-void rxr_pkt_handle_rtm_rta_recv(struct efa_rdm_ep *ep,
-				 struct efa_rdm_pke *pkt_entry);
-struct efa_rdm_ope *rxr_pkt_get_msgrtm_rxe(struct efa_rdm_ep *ep,
-						 struct efa_rdm_pke **pkt_entry_ptr);
-
-struct efa_rdm_ope *rxr_pkt_get_tagrtm_rxe(struct efa_rdm_ep *ep,
-						 struct efa_rdm_pke **pkt_entry_ptr);
 #endif
